Since Facebook Graph API v2.4 you need to specify which fields you want to receive.
https://developers.facebook.com/bugs/298946933534016/

(...) In Graph API v2.4 or higher, you must explicitly ask for the
'email' field of the user object (e.g. /v2.5/me?fields=email) even if
the 'email' permission scope is granted, and the user has a reachable
email address on file - this is due to the "Declarative Fields' entry in
the changelog [https://developers.facebook.com/docs/apps/changelog] and
applies to most fields of most objects in the API (...)

Even with the scope 'email' you just receive the 'id' and
'name'. There is no 'email'.

Just test it here:

https://developers.facebook.com/tools/explorer?method=GET&path=me&version=v2.6

With this modification you can get the old behaviour specifing the
fields
on the config file 'config.yml'. Even better! Add 'picture' to get the
profile/avatar photo and 'token_for_business' to identify the same user
across several apps/web belonging to you.

Example to replicate the behaviour from v2.0:

plugins:
    "Auth::OAuth"
        providers:
            Facebook:
                tokens:
                    client_id: 1111111111111111
                    client_secret: aaaaaaaaaaaaaaaaaaaaaa
                fields: email,first_name,gender,id,last_name,link,locale,middle_name,name,timezone,updated_time,verified

Or a simpler one assuming you are using the same user database for
several sites or apps:

    fields: token_for_business,email,id,name,gender,picture